# Get back performance in compressed pickle dumping and loading