--- a/joblib/numpy_pickle.py
+++ b/joblib/numpy_pickle.py
@@ -406,7 +406,6 @@ class ZipNumpyUnpickler(NumpyUnpickler):
         return BytesIO(read_zfile(file_handle))
 
 
-
 ###############################################################################
 # Utility functions