Fixes #194. Not sure this is the best way to do it so comments more than welcome.

As noted in the issue the numpy pickles are 1.5 times bigger for both
    - python 3.4 protocol=2 compared to python 3.4 protocol=4
    - python 3.4 protocol=2 compared to python 2.7 protocol=2

The reason for this is that for protocol < 3, save_bytes ends up creating a str [here](https://hg.python.org/releasing/3.4.3/file/933c3041d548/Lib/pickle.py#l686) which ends up being encoded in utf8 [here](https://hg.python.org/releasing/3.4.3/file/933c3041d548/Lib/pickle.py#l686).

I had to NumpyPickler.save_bytes override Pickler.save_bytes to work around this inefficiency and now the benchmarks are similar between joblib master and joblib 0.8.4.

The same benchmark script used in the issue output now looks like:

```
Python: 3.4
================================================================================
joblib file: /volatile/le243287/miniconda3/lib/python3.4/site-packages/sklearn/externals/joblib/__init__.py
--------------------
big arrays
Dumping took 4.5 s
Loading took 0.7 s
--------------------
big dict
Dumping took 11.3 s
Loading took 4.5 s
================================================================================
joblib file: /home/le243287/dev/joblib/joblib/__init__.py
--------------------
big arrays
Dumping took 4.3 s
Loading took 0.7 s
--------------------
big dict
Dumping took 11.4 s
Loading took 4.7 s
```