@ogrisel 2015-03-17
I think the backward compat on `unicode` and `bytes` strings (outside of numpy) should be more thoroughly tested to ensure that this does not introduce any regression.