@lesteve 2015-03-17
It looks like my solution was too naive. I get a str vs bytes mismatch in the test. Investigating.