@lesteve 2015-03-19
OK the tests are failing with Python 3.2 but once #196 is merged I'll rebase on master. Tests pass locally with Python 3.3 for me.

Here is a summary of this PR code changes which hopefully helps understanding them a bit more:
* for efficiency reasons we implemented save_bytes in the NumpyPickler to be identical to python 2.7 pickle.Pickler.save_bytes, which means that it is using SHORT_BINSTRING and BINSTRING markers and using the bytes directly rather than ending up decoding the bytes in utf-8 somewhere along the pickling code path for Python 3 and protocol=2.
* when we unpickle we don't want to do any decoding of the bytes between these markers because the numpy array can be created directly from them so encoding='bytes' makes sense
* encoding='bytes' does not exist for Python 3.3 so I ended up copying 4 functions from Python 3.4 in order to support this functionality

Any comments let me know.