@lesteve 2015-03-27
I rebased on master to drop python 3.2 support and get the 0.9.0.dev0 version and now tests pass fine as expected. Any comments about the meat of the changes, let me know.