@ogrisel 2015-04-08
Apart from this, this looks good to me.