@lesteve 2015-04-16
I added a description of the changes in CHANGES.rst.

@ogrisel, any further comments, in particular on the memory copy thing above ?