@ogrisel 2015-04-21
Thanks very much for this fix @lesteve. Merging.